==> RecSys2020_Challenge/01_Preprocess/GPU/2-preproc-split-new-DASK-MultiGPU.ipynb <==
Copyright (c) 2020, NVIDIA CORPORATION.
Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at
    http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

import os, time
os.environ["CUDA_VISIBLE_DEVICES"]="0,1,2,3"
start = time.time()
very_start = time.time()

#import pandas as pd, 
import numpy as np
from datetime import datetime
import matplotlib.pyplot as plt
#pd.set_option('display.max_columns', 500)
#pd.set_option('display.max_rows', 500)
import cudf, cupy, time, rmm
cudf.__version__

import dask as dask, dask_cudf
from dask.distributed import Client, wait
from dask_cuda import LocalCUDACluster
import subprocess

cluster = LocalCUDACluster(ip='10.2.61.36')#,protocol="ucx", 
                           #rmm_pool_size="26GB",
                           #enable_tcp_over_ucx=True, enable_nvlink=True)
#cluster = LocalCUDACluster()
client = Client(cluster)
client

%%time
path = '/raid/data/recsys/dask_input/step1_output'
df = dask_cudf.read_parquet(f'{path}/*.parquet')
df, = dask.persist(df)
_ = wait(df)
df.head()

%%time
def count_token(ds,token):
    not_null = ds.isnull()==0
    return ((ds.str.count(token)+1)*not_null).fillna(0)
df['len_hashtags'] = df['hashtags'].map_partitions(lambda ds: count_token(ds,'\t'))
df['len_domains']  = df['domains'].map_partitions(lambda ds: count_token(ds,'\t'))
df['len_links']    = df['links'].map_partitions(lambda ds: count_token(ds,'\t'))

df, = dask.persist(df)
_ = wait(df)
df.head()

def most_frequent_token(df, col, idcol, outcol, sep='\t'):
    ds = df[[idcol,col]].compute()
    tokens = ds[col].str.tokenize(sep)
    
    token_counts = ds[col].str.token_count('\t').to_frame()
    token_counts[idcol] = ds[idcol]
    del ds
    token_rows = reconstruct_token_rows(token_counts,col,idcol)
    del token_counts
    
    tokens = tokens.to_frame()
    tokens.columns = ['token']
    tokens['token'],_ = tokens['token'].factorize()
    tokens['token'] = tokens['token']+1
    tokens[idcol] = token_rows
    
    global_token_counts = tokens['token'].value_counts().reset_index()
    global_token_counts.columns = ['token','count']
    
    tokens = tokens.merge(global_token_counts,on='token',how='left')
    del global_token_counts
    
    token_counts = tokens.groupby(idcol).agg({'count':'max'}).reset_index()
    tokens = tokens.merge(token_counts,on=idcol,how='left')
    del token_counts
    mask = tokens.count_x == tokens.count_y
    tokens = tokens[mask]
    tokens = tokens.drop(['count_x','count_y'],axis=1)
    tokens = tokens.drop_duplicates(subset=['token',idcol])
    tokens = tokens.rename(mapper={'token':outcol})
    
    if outcol == col:
        df = df.drop(col,axis=1)
    df = df.merge(tokens,on=idcol,how='left')
    df[outcol] = df[outcol].fillna(0)
    del tokens
    return df

def reconstruct_token_rows(token_counts,col,idcol):
    ids = cupy.asnumpy(token_counts[idcol].values)
    rep = cupy.asnumpy(token_counts[col].values)
    token_rows = np.repeat(ids,rep)
    token_rows = cupy.asarray(token_rows)
    return token_rows


%%time
for col in ['domains','links','hashtags']:
    df = most_frequent_token(df, col=col, idcol='id', outcol=col, sep='\t')
    df, = dask.persist(df)
    _ = wait(df)
df.head()

%%time

TRAIN_SIZE = 121386431
VALID_SIZE = 12434735

train = df[df['id']<=TRAIN_SIZE] # df['id'].min() = 1
test0 = df[(df['id']>TRAIN_SIZE)&(df['id']<=TRAIN_SIZE+VALID_SIZE)]
test1 = df[df['id']>TRAIN_SIZE+VALID_SIZE]

print(len(train),len(test0),len(test1))

%%time
path = '/raid/data/recsys/dask_input/step2_output'

%%time
train.to_parquet(f'{path}/train.parquet', write_index=False)
test0.to_parquet(f'{path}/valid.parquet', write_index=False)
test1.to_parquet(f'{path}/test.parquet', write_index=False)

print('This notebook took %.1f minutes'%((time.time()-very_start)/60.))
